# Remove extra copy of entire file for Parquet reading