--- a/src/ArrowFunctions.cpp
+++ b/src/ArrowFunctions.cpp
@@ -106,7 +106,7 @@ int cpp_getType(const char* filename, const char* colname, char** errMsg) {
     return ARROWUNDEFINED;
 }
 
-int cpp_getStringColumnNumBytes(const char* filename, const char* colname, void* chpl_offsets, char** errMsg) {
+int cpp_getStringColumnNumBytes(const char* filename, const char* colname, void* chpl_offsets, int64_t numElems, int64_t startIdx, char** errMsg) {
   int64_t ty = cpp_getType(filename, colname, errMsg);
   auto offsets = (int64_t*)chpl_offsets;
 
@@ -139,12 +139,15 @@ int cpp_getStringColumnNumBytes(const char* filename, const char* colname, void*
       column_reader = row_group_reader->Column(idx);
       parquet::ByteArrayReader* ba_reader =
           static_cast<parquet::ByteArrayReader*>(column_reader.get());
+      ba_reader -> Skip(startIdx);
 
-      while (ba_reader->HasNext()) {
+      int64_t numRead = 0;
+      while (ba_reader->HasNext() && numRead < numElems) {
         parquet::ByteArray value;
         (void)ba_reader->ReadBatch(1, nullptr, nullptr, &value, &values_read);
         size += value.len + 1; // add 1 to account for zero that will be added in Chapel array
         offsets[i++] = value.len + 1;
+        numRead += values_read;
       }
     }
     return size;
@@ -152,7 +155,7 @@ int cpp_getStringColumnNumBytes(const char* filename, const char* colname, void*
   return ARROWUNDEFINED;
 }
 
-int cpp_readColumnByName(const char* filename, void* chpl_arr, const char* colname, int64_t numElems, int64_t batchSize, char** errMsg) {
+int cpp_readColumnByName(const char* filename, void* chpl_arr, const char* colname, int64_t numElems, int64_t startIdx, int64_t batchSize, char** errMsg) {
   int64_t ty = cpp_getType(filename, colname, errMsg);
   
   std::unique_ptr<parquet::ParquetFileReader> parquet_reader =
@@ -188,8 +191,9 @@ int cpp_readColumnByName(const char* filename, void* chpl_arr, const char* colna
       auto chpl_ptr = (int64_t*)chpl_arr;
       parquet::Int64Reader* reader =
         static_cast<parquet::Int64Reader*>(column_reader.get());
+      reader->Skip(startIdx);
 
-      while (reader->HasNext()) {
+      while (reader->HasNext() && i < numElems) {
         (void)reader->ReadBatch(batchSize, nullptr, nullptr, &chpl_ptr[i], &values_read);
         i+=values_read;
       }
@@ -197,9 +201,10 @@ int cpp_readColumnByName(const char* filename, void* chpl_arr, const char* colna
       auto chpl_ptr = (int64_t*)chpl_arr;
       parquet::Int32Reader* reader =
         static_cast<parquet::Int32Reader*>(column_reader.get());
+      reader->Skip(startIdx);
 
       int32_t* tmpArr = (int32_t*)malloc(batchSize * sizeof(int32_t));
-      while (reader->HasNext()) {
+      while (reader->HasNext() && i < numElems) {
         // Can't read directly into chpl_ptr because it is int64
         (void)reader->ReadBatch(batchSize, nullptr, nullptr, tmpArr, &values_read);
         for (int64_t j = 0; j < values_read; j++)
@@ -211,6 +216,7 @@ int cpp_readColumnByName(const char* filename, void* chpl_arr, const char* colna
       auto chpl_ptr = (bool*)chpl_arr;
       parquet::BoolReader* reader =
         static_cast<parquet::BoolReader*>(column_reader.get());
+      reader->Skip(startIdx);
 
       while (reader->HasNext()) {
         (void)reader->ReadBatch(batchSize, nullptr, nullptr, &chpl_ptr[i], &values_read);
@@ -218,12 +224,13 @@ int cpp_readColumnByName(const char* filename, void* chpl_arr, const char* colna
       }
     } else if(ty == ARROWSTRING) {
       auto chpl_ptr = (unsigned char*)chpl_arr;
-      parquet::ByteArrayReader* ba_reader =
+      parquet::ByteArrayReader* reader =
         static_cast<parquet::ByteArrayReader*>(column_reader.get());
+      reader->Skip(startIdx);
 
-      while (ba_reader->HasNext()) {
+      while (reader->HasNext() && i < numElems) {
         parquet::ByteArray value;
-        (void)ba_reader->ReadBatch(1, nullptr, nullptr, &value, &values_read);
+        (void)reader->ReadBatch(1, nullptr, nullptr, &value, &values_read);
         for(int j = 0; j < value.len; j++) {
           chpl_ptr[i] = value.ptr[j];
           i++;
@@ -234,9 +241,10 @@ int cpp_readColumnByName(const char* filename, void* chpl_arr, const char* colna
       auto chpl_ptr = (double*)chpl_arr;
       parquet::FloatReader* reader =
         static_cast<parquet::FloatReader*>(column_reader.get());
+      reader->Skip(startIdx);
 
       float* tmpArr = (float*)malloc(batchSize * sizeof(float));
-      while (reader->HasNext()) {
+      while (reader->HasNext() && i < numElems) {
         // Can't read directly into chpl_ptr because it is a double
         (void)reader->ReadBatch(batchSize, nullptr, nullptr, tmpArr, &values_read);
         for (int64_t j = 0; j < values_read; j++)
@@ -248,8 +256,9 @@ int cpp_readColumnByName(const char* filename, void* chpl_arr, const char* colna
       auto chpl_ptr = (double*)chpl_arr;
       parquet::DoubleReader* reader =
         static_cast<parquet::DoubleReader*>(column_reader.get());
+      reader->Skip(startIdx);
 
-      while (reader->HasNext()) {
+      while (reader->HasNext() && i < numElems) {
         (void)reader->ReadBatch(batchSize, nullptr, nullptr, &chpl_ptr[i], &values_read);
         i+=values_read;
       }
@@ -405,8 +414,8 @@ extern "C" {
     return cpp_getNumRows(chpl_str, errMsg);
   }
 
-  int c_readColumnByName(const char* filename, void* chpl_arr, const char* colname, int64_t numElems, int64_t batchSize, char** errMsg) {
-    return cpp_readColumnByName(filename, chpl_arr, colname, numElems, batchSize, errMsg);
+  int c_readColumnByName(const char* filename, void* chpl_arr, const char* colname, int64_t numElems, int64_t startIdx, int64_t batchSize, char** errMsg) {
+    return cpp_readColumnByName(filename, chpl_arr, colname, numElems, startIdx, batchSize, errMsg);
   }
 
   int c_getType(const char* filename, const char* colname, char** errMsg) {
@@ -422,8 +431,8 @@ extern "C" {
                                     errMsg);
   }
 
-  int c_getStringColumnNumBytes(const char* filename, const char* colname, void* chpl_offsets, char** errMsg) {
-    return cpp_getStringColumnNumBytes(filename, colname, chpl_offsets, errMsg);
+  int c_getStringColumnNumBytes(const char* filename, const char* colname, void* chpl_offsets, int64_t numElems, int64_t startIdx, char** errMsg) {
+    return cpp_getStringColumnNumBytes(filename, colname, chpl_offsets, numElems, startIdx, errMsg);
   }
 
   const char* c_getVersionInfo(void) {