--- a/src/ParquetMsg.chpl
+++ b/src/ParquetMsg.chpl
@@ -98,24 +98,24 @@ module ParquetMsg {
   }
 
   proc readFilesByName(A: [] ?t, filenames: [] string, sizes: [] int, dsetname: string, ty) throws {
-    extern proc c_readColumnByName(filename, chpl_arr, colNum, numElems, batchSize, errMsg): int;
+    extern proc c_readColumnByName(filename, chpl_arr, colNum, numElems, startIdx, batchSize, errMsg): int;
     var (subdoms, length) = getSubdomains(sizes);
+    var fileOffsets = (+ scan sizes) - sizes;
     
     coforall loc in A.targetLocales() do on loc {
       var locFiles = filenames;
       var locFiledoms = subdoms;
-      forall (filedom, filename) in zip(locFiledoms, locFiles) {
+      var locOffsets = fileOffsets;
+      forall (off, filedom, filename) in zip(locOffsets, locFiledoms, locFiles) {
         for locdom in A.localSubdomains() {
           const intersection = domain_intersection(locdom, filedom);
           if intersection.size > 0 {
             var pqErr = new parquetErrorMsg();
-            var col: [filedom] t;
-            if c_readColumnByName(filename.localize().c_str(), c_ptrTo(col),
-                                  dsetname.localize().c_str(), filedom.size, batchSize,
-                                  c_ptrTo(pqErr.errMsg)) == ARROWERROR {
+            if c_readColumnByName(filename.localize().c_str(), c_ptrTo(A[intersection.low]),
+                                  dsetname.localize().c_str(), intersection.size, intersection.low - off,
+                                  batchSize, c_ptrTo(pqErr.errMsg)) == ARROWERROR {
               pqErr.parquetError(getLineNumber(), getRoutineName(), getModuleName());
             }
-            A[filedom] = col;
           }
         }
       }
@@ -124,31 +124,33 @@ module ParquetMsg {
 
   proc calcSizesAndOffset(offsets: [] ?t, byteSizes: [] int, filenames: [] string, sizes: [] int, dsetname: string) throws {
     var (subdoms, length) = getSubdomains(sizes);
+    var fileOffsets = (+ scan sizes) - sizes;
     
     coforall loc in offsets.targetLocales() do on loc {
       var locFiles = filenames;
       var locFiledoms = subdoms;
-      forall (i, filedom, filename) in zip(sizes.domain, locFiledoms, locFiles) {
+      var locOffsets = fileOffsets;
+      forall (i, off, filedom, filename) in zip(sizes.domain, locOffsets, locFiledoms, locFiles) {
         for locdom in offsets.localSubdomains() {
           const intersection = domain_intersection(locdom, filedom);
           if intersection.size > 0 {
             var pqErr = new parquetErrorMsg();
-            var col: [filedom] t;
-            byteSizes[i] = getStrColSize(filename, dsetname, col);
-            offsets[filedom] = col;
+            byteSizes[i] = getStrColSize(filename, dsetname, offsets, intersection.size,
+                                         intersection.low, intersection.low - off);
           }
         }
       }
     }
   }
 
-  proc getStrColSize(filename: string, dsetname: string, offsets: [] int) throws {
-    extern proc c_getStringColumnNumBytes(filename, colname, offsets, errMsg): int;
+  proc getStrColSize(filename: string, dsetname: string, offsets: [] int, numElems: int, chplStartIdx: int, cStartIdx: int) throws {
+    extern proc c_getStringColumnNumBytes(filename, colname, offsets, numElems, startIdx, errMsg): int;
     var pqErr = new parquetErrorMsg();
     
     var byteSize = c_getStringColumnNumBytes(filename.localize().c_str(),
                                              dsetname.localize().c_str(),
-                                             c_ptrTo(offsets),
+                                             c_ptrTo(offsets[chplStartIdx]),
+                                             numElems, cStartIdx,
                                              c_ptrTo(pqErr.errMsg));
     if byteSize == ARROWERROR then
       pqErr.parquetError(getLineNumber(), getRoutineName(), getModuleName());