Previously, Parquet was having each node read the entire file of
any file that had an intersection with its local domain into a
temporary array and then assigning that local array to the
distributed array, which was resulting in extra communication
as well as keeping around a copy of the entire file, which would
mean that memory footprints on each node were much higher than
they needed to be.

This PR switches that to instead read directly into the local
portion of the distributed array and only the necessary number
of elements, skipping to the offset location of each file.

Here are some performance numbers collected on a 16-node-cs-hdr
using the `parquetIO` benchmark:

locales | master | this branch
-- | -- | --
1 | 2.18 GiB/s | 2.34 GiB/s
2 | 4.04 GiB/s | 4.71 GiB/s
4 | 7.45 GiB/s | 8.80 GiB/s
8 | 14.62 GiB/s | 16.67 GiB/s
16 | 27.67 GiB/s | 32.50 GiB/s

String reading: 
locales | master | this branch
-- | -- | --
1 | 1.34 GiB/s | 2.53 GiB/s
2 | 1.61 GiB/s | 2.77 GiB/s
4 | 1.75 GiB/s | 2.97 GiB/s
8 | 1.85 GiB/s | 2.71 GiB/s
16 | 1.86 GiB/s | 3.01 GiB/s