@bmcdonald3 2022-03-04
Converting this to a draft as I discovered a correctness error with this branch when running this test on multiple locales with string reading:

<details>

```python
import sys
import arkouda as ak

import pyarrow as pa
import pyarrow.parquet as pq
import numpy as np
import pandas as pd

ak.connect(server=sys.argv[1], port=sys.argv[2])

a = ak.random_strings_uniform(minlen=1, maxlen=8, size=100)
df = pd.DataFrame({'array': a.to_ndarray()})
table = pa.Table.from_pandas(df)
pq.write_table(table, "/lus/scratch/flash/mcdonald/strparquet-correct/test-filea")
b = ak.random_strings_uniform(minlen=1, maxlen=8, size=123)
df = pd.DataFrame({'array': b.to_ndarray()})
table = pa.Table.from_pandas(df)
pq.write_table(table, "/lus/scratch/flash/mcdonald/strparquet-correct/test-fileb")
c = ak.random_strings_uniform(minlen=1, maxlen=8, size=64)
df = pd.DataFrame({'array': c.to_ndarray()})
table = pa.Table.from_pandas(df)
pq.write_table(table, "/lus/scratch/flash/mcdonald/strparquet-correct/test-filec")
d = ak.random_strings_uniform(minlen=1, maxlen=8, size=689)
df = pd.DataFrame({'array': d.to_ndarray()})
table = pa.Table.from_pandas(df)
pq.write_table(table, "/lus/scratch/flash/mcdonald/strparquet-correct/test-filed")
e = ak.random_strings_uniform(minlen=1, maxlen=8, size=23)
df = pd.DataFrame({'array': e.to_ndarray()})
table = pa.Table.from_pandas(df)
pq.write_table(table, "/lus/scratch/flash/mcdonald/strparquet-correct/test-filee")
f = ak.random_strings_uniform(minlen=1, maxlen=8, size=199)
df = pd.DataFrame({'array': f.to_ndarray()})
table = pa.Table.from_pandas(df)
pq.write_table(table, "/lus/scratch/flash/mcdonald/strparquet-correct/test-filef")

ak_vals = ak.read_parquet("/lus/scratch/flash/mcdonald/strparquet-correct/test-file*")

assert((a == ak_vals[0:100]).all())
assert((b == ak_vals[100:223]).all())
assert((c == ak_vals[223:287]).all())
for i in range(len(d)):
    if d[i] != ak_vals[i+287]:
        print(d[i], " does not equal ", ak_vals[i+287])
assert((d == ak_vals[287:976]).all())
assert((e == ak_vals[976:999]).all())
assert((f == ak_vals[999:1198]).all())
```

</details>